Fixed bugs where scraping would stop due to error.
Changed the saving of images to have a library do it for me. Too many cases where an image would just not save for some reason and puppeteer was just a bad choice for saving in the first place. I think I will keep the header though because that helps me with naming the file.